select according to size/age ratio

+++ src/settings.ml
@@ -34,7 +34,7 @@ type t = {
  pcp : int ref;
  extended_signature: bool ref;
  keep_orientation: bool ref;
- size_age_ratio: float ref
+ size_age_ratio: int ref
 }
 
 (*** GLOBALS *****************************************************************)
@@ -62,7 +62,7 @@ let default = {
  pcp = ref 0;
  extended_signature = ref false;
  keep_orientation = ref false;
- size_age_ratio = ref 100.
+ size_age_ratio = ref 100
 }
 
 let do_assertions = ref false
